requirements.txt
mcp
rss-parser
requests
beautifulsoup4
python-dateutil
lxml